relocated navigation for more flexibility
