+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import './App.css'
 import { Outlet } from 'react-router-dom'
-import Navigation from './pages/Navigation'
+import Navigation from './components/Navigation'
 
 function App() {
     const NavOptions = {
@@ -10,7 +10,6 @@ function App() {
 
     return (
         <div className="App">
-            <Navigation />
             <Outlet />
         </div>
     )
