+++ src/pages/Confirmation.tsx
@@ -1,4 +1,5 @@
 import { Link, useLocation } from 'react-router-dom'
+import Navigation from '../components/Navigation'
 interface confirmationState {
     header: string
     message: string
@@ -8,7 +9,8 @@ interface confirmationState {
 
 function Confirmation() {
     const state: confirmationState = useLocation().state
-    return (
+    return (<>
+        <Navigation />
         <div className="confirmation">
             <div className="confirmationBody">
                 <h1>{state.header}</h1>
@@ -18,7 +20,7 @@ function Confirmation() {
                 </Link>
             </div>
         </div>
-    )
+    </>)
 }
 
 export default Confirmation
