+++ src/pages/Landing.tsx
@@ -3,6 +3,7 @@ import speedyLogo from '../assets/speedyLogo.png'
 import { useState } from 'react'
 import { Link } from 'react-router-dom'
 import { landingInformationCards } from '../utils/dataObjects'
+import Navigation from '../components/Navigation'
 function Landing() {
     const [currentActivePanel, setCurrentActivePanel] = useState('about')
     const changeFocus = ({ innerText }: { innerText: string }) => {
@@ -10,43 +11,46 @@ function Landing() {
     }
 
     return (
-        <div className="landing">
-            <h1>
-                Revolutionize Workforce Management with
-                <br />
-                <span className="title">Speedy Check-In</span>
-            </h1>
+        <>
+            <Navigation />
+            <div className="landing">
+                <h1>
+                    Revolutionize Workforce Management with
+                    <br />
+                    <span className="title">Speedy Check-In</span>
+                </h1>
 
-            <div className="landingDescription">
-                <div className="logoSection">
-                    <img src={speedyLogo} alt="Speedy Checkin Logo" />
-                </div>
-                <div className="greeting">
-                    <p>
-                        Welcome to Speedy Check-In, your comprehensive solution
-                        for efficient workforce management!
-                    </p>
-                    <div className="actionLinks">
-                        <Link to="/login">
-                            <button>Login</button>
-                        </Link>
-                        <Link to={'/signup'}>
-                            <button>Sign-Up</button>
-                        </Link>
+                <div className="landingDescription">
+                    <div className="logoSection">
+                        <img src={speedyLogo} alt="Speedy Checkin Logo" />
                     </div>
-                </div>
+                    <div className="greeting">
+                        <p>
+                            Welcome to Speedy Check-In, your comprehensive
+                            solution for efficient workforce management!
+                        </p>
+                        <div className="actionLinks">
+                            <Link to="/login">
+                                <button>Login</button>
+                            </Link>
+                            <Link to={'/signup'}>
+                                <button>Sign-Up</button>
+                            </Link>
+                        </div>
+                    </div>
 
-                {landingInformationCards.map((card) => (
-                    <LandingInformationCard
-                        key={card.panelKey}
-                        cardTitle={card.cardTitle}
-                        cardContent={card.cardContent}
-                        isActive={currentActivePanel === card.panelKey}
-                        changeFocus={changeFocus}
-                    />
-                ))}
+                    {landingInformationCards.map((card) => (
+                        <LandingInformationCard
+                            key={card.panelKey}
+                            cardTitle={card.cardTitle}
+                            cardContent={card.cardContent}
+                            isActive={currentActivePanel === card.panelKey}
+                            changeFocus={changeFocus}
+                        />
+                    ))}
+                </div>
             </div>
-        </div>
+        </>
     )
 }
